feature:[a3d]linux编译问题

+++ cxnd/algrithm/a3d.cpp
@@ -233,10 +233,10 @@ namespace cxnd
 
 	trimesh::box3 extendBoxXY(const trimesh::box3& box, const trimesh::vec3& center)
 	{
-		float xmax = std::max(std::fabsf(box.max.x - center.x), 
-							std::fabsf(box.min.x - center.x));
-		float ymax = std::max(std::fabsf(box.max.y - center.y),
-							std::fabsf(box.min.y - center.y));
+		float xmax = std::max(std::fabs(box.max.x - center.x),
+							std::fabs(box.min.x - center.x));
+		float ymax = std::max(std::fabs(box.max.y - center.y),
+							std::fabs(box.min.y - center.y));
 
 		trimesh::vec3 dmin = center - trimesh::vec3(xmax, ymax, 0.0f);
 		dmin.z = box.min.z;
